refactor(core): promote pkg/portbind to internal/core/portbind (no behavior change)
